backend/app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
langdetect==1.0.9
pydantic==2.9.2
torch==2.2.2
huggingface_hub==0.19.4
sentence-transformers==2.2.2
numpy==1.24.3

